Consolidate bulbs-poll tests
