+++ elements/bulbs-poll/tests/bulbs-poll-integration.test.js
@@ -1,6 +1,6 @@
-//import { assert } from 'chai';
-//import fetchMock from 'fetch-mock';
-//import testElement from 'bulbs-elements/test/element';
+// import { assert } from 'chai';
+// import fetchMock from 'fetch-mock';
+// import testElement from 'bulbs-elements/test/element';
 // testElement('BulbsPoll', function () {
 //   beforeEach(function (done) {
 //     let pollEndpoint = 'http://example.tld/api/polls/1';
@@ -29,26 +29,26 @@
 //     });
 //     this.actions = this.element.reactElement.store.actions;
 //   });
-//
+
 //   it('renders a poll', function () {
 //     let answers = this.element.querySelectorAll('.bulbs-poll-answer');
 //     assert.equal(answers[0].textContent, 'answer A');
 //     assert.equal(answers[1].textContent, 'answer B');
 //   });
-//
+
 //   it('selects an answer', function () {
 //     let answer = this.element.querySelector('.bulbs-poll-answer');
 //     answer.click();
 //     assert.isTrue(answer.classList.contains('selected'));
 //   });
-//
+
 //   it('toggles an answer if clicked twice', function () {
 //     let answer = this.element.querySelector('.bulbs-poll-answer');
 //     answer.click();
 //     answer.click();
 //     assert.isFalse(answer.classList.contains('selected'));
 //   });
-//
+
 //   it('deselects an answer if a different anwser is selected', function () {
 //     let answers = this.element.querySelectorAll('.bulbs-poll-answer');
 //     answers[0].click();
@@ -56,14 +56,14 @@
 //     assert.isFalse(answers[0].classList.contains('selected'));
 //     assert.isTrue(answers[1].classList.contains('selected'));
 //   });
-//
+
 //   context('VoteButton', function () {
 //     it('will not vote when no answer is selected', function () {
 //       let actionSpy = chai.spy.on(this.actions, 'invoke');
 //       this.element.querySelector('.bulbs-poll-vote').click();
 //       actionSpy.should.not.have.been.called();
 //     });
-//
+
 //     it('will make a vote request if an answer is selected', function () {
 //       let actionSpy = chai.spy.on(this.actions, 'makeVoteRequest');
 //       this.element.querySelector('.bulbs-poll-answer').click();
@@ -71,10 +71,10 @@
 //       actionSpy.should.have.been.called();
 //     });
 //   });
-//
+
 //   context('voteRequestError', function () {
 //   });
-//
+
 //   context('voteRequestSuccess', function () {
 //     beforeEach(function () {
 //       let { voteRequestSuccess } = this.actions;
@@ -97,13 +97,13 @@
 //         },
 //       });
 //     });
-//
+
 //     it('does not render a vote button', function () {
 //       let voteButton = this.element.querySelector('button');
 //       assert.isNull(voteButton);
 //     });
 //   });
-//
+
 //   context('voteRequestFailure', function () {
 //     beforeEach(function () {
 //       let { voteRequestFailure } = this.actions;
@@ -117,13 +117,13 @@
 //         status: 400,
 //         json: () => {
 //           return {
-//
+
 //           };
 //         },
 //       });
 //     });
 //   });
-//
+
 //   context('voteRequestError', function () {
 //     it('displays an error message', function () {
 //       let { voteRequestError } = this.actions;
