+++ elements/bulbs-poll/tests/bulbs-poll.test.js
@@ -1,7 +1,6 @@
-//import { assert } from 'chai';
 import fetchMock from 'fetch-mock';
 import testElement from 'bulbs-elements/test/element';
-import './bulbs-poll';
+import '../bulbs-poll';
 
 testElement('<bulbs-poll> <BulbsPoll>', function () {
   let setSrcSpy;
